+++ modules/snipe/choice_type/__init__.py
@@ -1,6 +1,7 @@
 from .base import *
 from .emoji import *
 from .reaction import *
+from .soundboard_sound import *
 from .sticker import *
 
 
@@ -8,9 +9,11 @@ __all__ = (
     *base.__all__,
     *emoji.__all__,
     *reaction.__all__,
+    *soundboard_sound.__all__,
     *sticker.__all__
 )
 
 from .emoji import ChoiceTypeEmoji
 from .reaction import ChoiceTypeReaction
+from .soundboard_sound import ChoiceTypeSoundboardSound
 from .sticker import ChoiceTypeSticker