+++ modules/snipe/interaction_reveal.py
@@ -6,7 +6,7 @@ from bots import SLASH_CLIENT
 
 from .component_translate_tables import REVEAL_DISABLE
 from .constants import CUSTOM_ID_SNIPE_REVEAL
-from .helpers import translate_components
+from .helpers import get_message_attachment, translate_components
 
 
 @SLASH_CLIENT.interactions(custom_id = CUSTOM_ID_SNIPE_REVEAL)
@@ -32,9 +32,11 @@ async def snipe_message_reveal(client, event):
         return
     
     await client.interaction_component_acknowledge(event, wait = False)
+    file = await get_message_attachment(client, message)
     await client.interaction_followup_message_create(
         event,
         embed = embed,
         components = translate_components(event.message.iter_components(), REVEAL_DISABLE),
+        file = file,
     )
     await client.interaction_response_message_delete(event)